+++ src/main/ts/utils/NewOrCallUtils.ts
@@ -32,5 +32,5 @@ export const forIdentifier = (f: (node: CallExpression | NewExpression | Identif
         checkArguments(node);
       }
     }
-  }
+  };
 };