+++ CHANGELOG.md
@@ -8,8 +8,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 1.9.0 - 2021-06-14
 
+### Added
+- The `@tinymce/prefer-fun` rule now supports configuring which rules should be checked.
+
 ### Improved
 - The `@tinymce/no-main-module-imports` rule now detects Main imports via aliases from outside the `src/main/` directory.
+- The `@tinymce/prefer-fun` rule now detects where `Fun.constant` should be used.
 
 ### Changed
 - The editor configuration now enforces newlines between import groups.